make stuff public
